feat(ColumnResize): compute the width of colums automatically
This new feature enables ui-grid to automatically compute the width of columns, which do not have a specified width. The same algorithm used by auto resizing fired by double clicking is generalized and used to compute the width of newly added columns with undefined width. The computing of xDiff is furthermore corrected. A new option is also added to specify how many rows should be used to determine the auto width of columns.