fix(autoResize): fix the computation of width/height to enable scroll
This fix is only relevant for IE 11 and older versions, where gridUtil.outerElementHeight or outerElementWidth does not return correct values if ui-grid has padding or borders. This leads to the problem of not allowing the page to scroll, if the mouse is on the list, even if there is nothing to scroll down in ui-grid.

This solution uses jQuery and it cannot be submitted as a pull request according to the guidelines of ui-grid.

Alternatively, one needs to expand the methods in gridUtil to the case of IE11 and older.